+++ src/Core/Domain/Boss4D.Core.Domain.Package.pas
@@ -71,6 +71,7 @@ type
     FProjects: TList<string>;
     FScripts: TDictionary<string, string>;
     FDependencies: TDictionary<string, string>;
+    FDevDependencies: TDictionary<string, string>;
     FEngines: TBoss4DPackageEngines;
     FToolchain: TBoss4DPackageToolchain;
     FWorkspaces: TList<string>;
@@ -80,11 +81,13 @@ type
     destructor Destroy; override;
 
     procedure AddDependency(const ADep: string; const AVer: string);
+    procedure AddDevDependency(const ADep: string; const AVer: string);
     function RemoveDependency(const ADep: string): Boolean;
     procedure AddProject(const AProject: string);
 
     // Retorna uma lista de dependencias prontas e parseadas. O chamador e responsavel por liberar os objetos do array.
     function GetParsedDependencies: TArray<TBoss4DDependency>;
+    function GetParsedDevDependencies: TArray<TBoss4DDependency>;
 
     property Name: string read FName write FName;
     property Description: string read FDescription write FDescription;
@@ -96,6 +99,7 @@ type
     property Projects: TList<string> read FProjects;
     property Scripts: TDictionary<string, string> read FScripts;
     property Dependencies: TDictionary<string, string> read FDependencies;
+    property DevDependencies: TDictionary<string, string> read FDevDependencies;
     property Engines: TBoss4DPackageEngines read FEngines;
     property Toolchain: TBoss4DPackageToolchain read FToolchain;
     property Workspaces: TList<string> read FWorkspaces;
@@ -129,6 +133,7 @@ begin
   FProjects := TList<string>.Create;
   FScripts := TDictionary<string, string>.Create;
   FDependencies := TDictionary<string, string>.Create;
+  FDevDependencies := TDictionary<string, string>.Create;
   FEngines := TBoss4DPackageEngines.Create;
   FToolchain := TBoss4DPackageToolchain.Create;
   FWorkspaces := TList<string>.Create;
@@ -141,6 +146,7 @@ begin
   FWorkspaces.Free;
   FToolchain.Free;
   FEngines.Free;
+  FDevDependencies.Free;
   FDependencies.Free;
   FScripts.Free;
   FProjects.Free;
@@ -164,6 +170,23 @@ begin
     FDependencies.AddOrSetValue(LFoundKey, AVer)
   else
     FDependencies.Add(ADep, AVer);
+end;
+
+procedure TBoss4DPackage.AddDevDependency(const ADep, AVer: string);
+var
+  LKey, LFoundKey: string;
+begin
+  LFoundKey := '';
+  for LKey in FDevDependencies.Keys do
+    if SameText(LKey, ADep) then
+    begin
+      LFoundKey := LKey;
+      Break;
+    end;
+  if not LFoundKey.IsEmpty then
+    FDevDependencies.AddOrSetValue(LFoundKey, AVer)
+  else
+    FDevDependencies.Add(ADep, AVer);
 end;
 
 function TBoss4DPackage.RemoveDependency(const ADep: string): Boolean;
@@ -200,6 +223,27 @@ begin
     for var LPair in FDependencies do
     begin
       LResultList.Add(TBoss4DDependency.Parse(LPair.Key, LPair.Value));
+    end;
+    Result := LResultList.ToArray;
+  finally
+    LResultList.Free;
+  end;
+end;
+
+function TBoss4DPackage.GetParsedDevDependencies: TArray<TBoss4DDependency>;
+var
+  LResultList: TList<TBoss4DDependency>;
+  LDep: TBoss4DDependency;
+begin
+  if FDevDependencies.Count = 0 then
+    Exit(nil);
+  LResultList := TList<TBoss4DDependency>.Create;
+  try
+    for var LPair in FDevDependencies do
+    begin
+      LDep := TBoss4DDependency.Parse(LPair.Key, LPair.Value);
+      LDep.Scope := 'development';
+      LResultList.Add(LDep);
     end;
     Result := LResultList.ToArray;
   finally